+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - [Catalyst Engines demo](https://github.com/catalyst-team/catalyst/tree/master/examples/engines)
 - `dataset_from_params` support in config API ([#1231](https://github.com/catalyst-team/catalyst/pull/1231))
 - transform from params support for config API added ([#1236](https://github.com/catalyst-team/catalyst/pull/1236))
+- samplers from params support for config API added ([#1240](https://github.com/catalyst-team/catalyst/pull/1240))
 
 ### Changed